# swagger.json
{"swagger":"2.0","info":{"description":"This is the API Documentation for More-Recipes. More-Recipes is a recipe sharing and social connecting platform where users can show their cuisine expertise by posting recipes for other users to view and learn from","version":"1.0.0","title":"More-Recipes API","contact":{"email":"emperoarkay@gmail.com"},"license":{"name":"MIT","url":"https://github.com/swagger-api/swagger-ui/blob/master/LICENSE"}},"host":"more-recipes17.herokuapp.com","basePath":"/api/v1","tags":[{"name":"users","description":"Users related operations endpoints"},{"name":"recipes","description":"Recipes related operations endpoints"}],"schemes":["https"],"paths":{"/users/signup":{"post":{"tags":["users"],"summary":"Create new user on the system","operationId":"createUser","description":"By filling in valid form inputs, a new user can create an account","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"description":"Form input model","schema":{"$ref":"#/definitions/User"}}],"responses":{"201":{"description":"Account successfully created","schema":{"$ref":"#/definitions/Signup"}},"400":{"description":"Validation exception","schema":{"$ref":"#/definitions/SignupError"}},"409":{"description":"A user with same username or email already exists","schema":{"$ref":"#/definitions/SignupValidation"}},"default":{"description":"Unexpected Error"}}}},"/users/signin":{"post":{"tags":["users"],"summary":"Authenticates and logs user into the system","operationId":"loginUser","description":"By providing valid login credentials a user is able to login to the system","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"description":"Login credentials","schema":{"$ref":"#/definitions/Login"}}],"responses":{"200":{"description":"Login successful","schema":{"$ref":"#/definitions/Signin"}},"401":{"description":"Invalid username or password","schema":{"$ref":"#/definitions/SigninError"}},"default":{"description":"Unexpected Error"}}}},"/users/changepassword":{"put":{"tags":["users"],"summary":"Allows an authenticated user change password","operationId":"changePassword","description":"By providing valid login credentials a user is able to change his/her password","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"in":"body","name":"password","required":true,"description":"New password to be changed to","schema":{"$ref":"#/definitions/ChangePassword"}}],"responses":{"200":{"description":"Password successfully changed","schema":{"$ref":"#/definitions/ChangePasswordResponse"}},"401":{"description":"Request could not be authorized","schema":{"$ref":"#/definitions/ChangePasswordError"}},"400":{"description":"Validation exception","schema":{"$ref":"#/definitions/ChangePasswordError"}},"default":{"description":"Unexpected Error"}}}},"/users/{recipeId}/favorites":{"post":{"tags":["users"],"summary":"Add a recipe to a user favorite list","operationId":"addFavorite","description":"API endpoint for adding a recipe as a users favorite","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"in":"path","name":"recipeId","required":true,"description":"ID of recipe to be added as favorite","type":"integer"},{"in":"body","name":"category","required":false,"description":"category favorited recipe should be categorized under","schema":{"$ref":"#/definitions/Category"}}],"responses":{"201":{"description":"Recipe successfully added to favorites","schema":{"$ref":"#/definitions/FavoriteResponse"}},"404":{"description":"Recipe not found","schema":{"$ref":"#/definitions/RecipeNotFound"}},"409":{"description":"Item already exists","schema":{"$ref":"#/definitions/FavoriteExists"}},"422":{"description":"Invalid recipeId parameter","schema":{"$ref":"#/definitions/InvalidParameter"}},"default":{"description":"Unexpected Error"}}},"patch":{"tags":["users"],"summary":"Modify or assign category to a favorite recipe on a user's favorite recipe list","operationId":"addRecipeCategory","description":"This endpoint allows logged in user to modify or assign a category to recipes on his/her favorite list","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"in":"path","name":"recipeId","required":true,"description":"ID of favorite recipe to be modified","type":"integer"},{"in":"body","name":"category","required":false,"description":"category favorited recipe should be categorized under","schema":{"$ref":"#/definitions/Category"}}],"responses":{"200":{"description":"Recipe added to category","schema":{"$ref":"#/definitions/CategoryResponse"}},"404":{"description":"Recipe not found","schema":{"$ref":"#/definitions/RecipeNotFound"}},"409":{"description":"Recipe has not been added to favorite","schema":{"$ref":"#/definitions/FavoriteException"}},"422":{"description":"Invalid recipeId parameter","schema":{"$ref":"#/definitions/InvalidParameter"}},"default":{"description":"Unexpected Error"}}},"delete":{"tags":["users"],"summary":"Remove a recipe from a user's favorite list","operationId":"deleteFavorite","description":"Method allows logged in user to remove recipes from his/her favorite list","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"in":"path","name":"recipeId","required":true,"description":"ID of recipe to be removed from favorite","type":"integer"}],"responses":{"200":{"description":"Operation successful","schema":{"$ref":"#/definitions/DeleteResponse"}},"default":{"description":"Unexpected Error"},"404":{"description":"Recipe not found","schema":{"$ref":"#/definitions/RecipeNotFound"}},"409":{"description":"Recipe not yet favorited","schema":{"$ref":"#/definitions/FavoriteException"}},"422":{"description":"Invalid recipeId parameter","schema":{"$ref":"#/definitions/InvalidParameter"}}}}},"/users/recipes/favorites":{"get":{"tags":["users"],"summary":"Fetch at most 10 recipes in user's favorite list","operationId":"getUserFavorites","description":"Method allows logged in user to view at most 10 recipes from his/her favorite list","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/FavoriteRecipes"}}},"default":{"description":"Unexpected error"}}}},"/recipes":{"post":{"tags":["recipes"],"summary":"Create and post a recipe","operationId":"createRecipePost","description":"API endpoint for creating a recipe","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"name":"body","in":"body","required":true,"schema":{"description":"Recipe model","$ref":"#/definitions/Recipe"}}],"responses":{"201":{"description":"Operation successful","schema":{"description":"Responds with a recipe instance","$ref":"#/definitions/CreatedRecipe"}},"400":{"description":"Validation exception","schema":{"$ref":"#/definitions/RecipeValidation"}},"default":{"description":"Unexpected error"}}},"get":{"tags":["recipes"],"summary":"Fetch at most 10 recipes in the application","operationId":"getRecipes","description":"Endpoint for fetching recipe posts","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/GetRecipe"}}},"default":{"description":"Unexpected Error"}}}},"/recipes?sort=upvote&order=descending":{"get":{"tags":["recipes"],"summary":"Fetch at most 5 recipes with the highest upvotes","operationId":"getTopRecipes","description":"API endpoint for fetching top recipes using most upvotes criteria","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/GenericRecipe"}}},"default":{"description":"Unexpected error"}}}},"/recipes?category=pastries":{"get":{"tags":["recipes"],"summary":"Search for recipes that match category query string","operationId":"searchRecipesByCategory","description":"API endpoint for searching recipe posts by category","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/GenericRecipe"}}},"default":{"description":"Unexpected error"}}}},"/recipes?ingredients=flour":{"get":{"tags":["recipes"],"summary":"Search for recipes that match ingredients query string","operationId":"searchRecipesByIngredients","description":"API endpoint for searching recipe posts that match ingredients value","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/GenericRecipe"}}},"default":{"description":"Unexpected error"}}}},"/recipes/{recipeId}":{"put":{"tags":["recipes"],"summary":"Modify a recipe","operationId":"modifyRecipePost","description":"Edit contents of a posted recipe","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"in":"path","name":"recipeId","required":true,"description":"ID of recipe to be modified","type":"integer"},{"in":"body","name":"body","description":"Recipe form model","schema":{"$ref":"#/definitions/ModifyRecipe"}}],"responses":{"200":{"description":"Modification successful","schema":{"$ref":"#/definitions/ModifiedRecipe"}},"422":{"description":"Invalid recipeId parameter","schema":{"$ref":"#/definitions/InvalidParameter"}},"default":{"description":"Unexpected error"}}},"get":{"tags":["recipes"],"summary":"View a recipe details","operationId":"viewRecipePost","description":"Retrieve a single recipe post","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"in":"path","name":"recipeId","required":true,"description":"ID of recipe","type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ViewRecipe"}},"422":{"description":"Invalid recipeId parameter","schema":{"$ref":"#/definitions/InvalidParameter"}},"default":{"description":"Unexpected error"}}},"delete":{"tags":["recipes"],"summary":"Delete Recipe","operationId":"deleteRecipePost","description":"Deletes a user's recipe post, action cannot be undone","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"in":"path","name":"recipeId","required":true,"description":"ID of recipe to be deleted","type":"integer"}],"responses":{"200":{"description":"Operation successful","schema":{"$ref":"#/definitions/DeleteRecipe"}},"404":{"description":"Not found","schema":{"$ref":"#/definitions/RecipeNotFound"}},"422":{"description":"Invalid recipeId","schema":{"$ref":"#/definitions/InvalidParameter"}},"default":{"description":"Unexpected error"}}}},"/recipes/user":{"get":{"tags":["recipes"],"summary":"View recipes posted by user","operationId":"getUserRecipes","description":"Fetch at most ten recipes posted by user on the system","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/GenericRecipe"}}},"default":{"description":"Unexpected Error"}}}},"/recipes/{recipeId}/reviews":{"post":{"tags":["recipes"],"summary":"Comment on a recipe post","operationId":"postReview","description":"Users review a post by commenting on it","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"in":"path","name":"recipeId","required":true,"description":"ID of recipe to be reviewed","type":"integer"},{"in":"body","name":"comment","required":true,"description":"comment by user","schema":{"$ref":"#/definitions/PostedReview"}}],"responses":{"200":{"description":"Comment posted","schema":{"$ref":"#/definitions/ReviewResponse"}},"404":{"description":"Not found","schema":{"$ref":"#/definitions/RecipeNotFound"}},"422":{"description":"Invalid recipeId","schema":{"$ref":"#/definitions/InvalidParameter"}},"default":{"description":"Unexpected error"}}}},"/recipes/{recipeId}/upvote":{"put":{"tags":["recipes"],"summary":"Upvote recipe post","operationId":"upvoteRecipe","description":"A user can upvote a recipe he/she likes","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"in":"path","name":"recipeId","required":true,"description":"ID of recipe being upvoted","type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UpvoteResponse"}},"422":{"description":"Bad input parameter","schema":{"$ref":"#/definitions/InvalidParameter"}},"404":{"description":"Not found","schema":{"$ref":"#/definitions/RecipeNotFound"}},"default":{"description":"Unexpected error"}}}},"/recipes/{recipeId}/downvote":{"put":{"tags":["recipes"],"summary":"Downvote recipe post","operationId":"downvoteRecipe","description":"A user can downvote a recipe","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"x-access-token","in":"header","description":"Authentication Token","required":true,"type":"string"},{"in":"path","name":"recipeId","required":true,"description":"ID of recipe being downvoted","type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DownvoteResponse"}},"422":{"description":"Bad input parameter","schema":{"$ref":"#/definitions/InvalidParameter"}},"404":{"description":"Not found","schema":{"$ref":"#/definitions/RecipeNotFound"}},"default":{"description":"Unexpected error"}}}}},"definitions":{"Recipe":{"type":"object","required":["recipeName","ingredients","instructions"],"properties":{"recipeName":{"type":"string","example":"cake"},"category":{"type":"string","example":"pastries"},"ingredients":{"type":"string","example":"baking powder, flour, margarine"},"instructions":{"type":"string","example":"pour and frisk egg with margarine"}}},"User":{"type":"object","required":["username","email","password","confirmPassword"],"properties":{"username":{"type":"string","example":"johndoe17"},"email":{"type":"string","example":"johndoe@company.com"},"password":{"type":"string","example":"password"},"confirmPassword":{"type":"string","example":"password"}}},"Login":{"type":"object","required":["username","password"],"properties":{"username":{"type":"string","example":"johndoe17"},"password":{"type":"string","example":"password"}}},"Signup":{"type":"object","required":["status","message","token"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Account successfully created"},"token":{"type":"string","example":"foo.bar.baz"}}},"SignupError":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"fail"},"message":{"type":"string","example":"Invalid Email, please enter a valid email"}}},"SignupValidation":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"fail"},"message":{"type":"string","example":"Username already exists"}}},"Signin":{"type":"object","required":["status","message","Token"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Token successfully generated"},"Token":{"type":"string","example":"foo.bar.baz"}}},"SigninError":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"fail"},"message":{"type":"string","example":"Invalid Authentication Details"}}},"ChangePassword":{"type":"object","required":["password"],"properties":{"password":{"type":"string","example":"newpassword"}}},"ChangePasswordResponse":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Password changed successfully"}}},"ChangePasswordError":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"fail"},"message":{"type":"string","example":"Password should be at least six characters long"}}},"Category":{"type":"object","properties":{"category":{"type":"string","example":"dessert"}}},"FavoriteRecipes":{"type":"object","required":["id","userId","recipeId"],"properties":{"id":{"type":"integer","format":"int64","example":2},"userId":{"type":"integer","format":"int64","example":4},"recipeId":{"type":"integer","format":"int64","example":6},"category":{"type":"string","example":"african dish"}}},"DeleteResponse":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Recipe successfully removed from favorites"}}},"RecipeNotFound":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"fail"},"message":{"type":"string","example":"Recipe not found"}}},"FavoriteException":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"fail"},"message":{"type":"string","example":"Recipe has not been added to favorite"}}},"InvalidParameter":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"fail"},"message":{"type":"string","example":"You have entered an invalid parameter"}}},"FavoriteResponse":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Recipe successfully added to favorites"}}},"FavoriteExists":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"fail"},"message":{"type":"string","example":"Recipe has already been favorited"}}},"CategoryResponse":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Recipe added to dessert"}}},"CreatedRecipe":{"type":"object","required":["status","message","views","upvote","downvote","recipeName","ingredients","instructions"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Successfully created new recipe"},"views":{"type":"integer","format":"int64","example":1},"upvote":{"type":"integer","format":"int64","example":0},"downvote":{"type":"integer","format":"int64","example":0},"recipeName":{"type":"string","example":"cake"},"category":{"type":"string","example":"pastries"},"ingredients":{"type":"string","example":"baking powder, flour, margarine"},"instructions":{"type":"string","example":"pour and frisk egg with margarine"}}},"RecipeValidation":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"fail"},"message":{"type":"string","example":"Please enter a recipe name"}}},"Comment":{"type":"object","required":["comment","User"],"properties":{"comment":{"type":"string","example":"comment from user"},"User":{"type":"object","required":["username","createdAt"],"properties":{"username":{"type":"string","example":"johndoe17"},"createdAt":{"type":"string","example":"2017-09-08T20:01:18.086Z"}}}}},"GetRecipe":{"type":"object","required":["id","views","upvote","downvote","recipeName","category","ingredients","instructions","reviews"],"properties":{"id":{"type":"integer","format":"int64","example":1},"views":{"type":"integer","format":"int64","example":15},"upvote":{"type":"integer","format":"int64","example":8},"downvote":{"type":"integer","format":"int64","example":2},"recipeName":{"type":"string","example":"cake"},"category":{"type":"string","example":"pastries"},"ingredients":{"type":"string","example":"baking powder, flour, margarine"},"instructions":{"type":"string","example":"pour and frisk egg with margarine"},"reviews":{"type":"array","items":{"$ref":"#/definitions/Comment"}}}},"ModifiedRecipe":{"type":"object","required":["status","message","views","upvote","downvote","recipeName","ingredients","instructions"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Recipe successfully updated"},"views":{"type":"integer","format":"int64","example":1},"upvote":{"type":"integer","format":"int64","example":0},"downvote":{"type":"integer","format":"int64","example":0},"recipeName":{"type":"string","example":"cake"},"category":{"type":"string","example":"pastries"},"ingredients":{"type":"string","example":"baking powder, flour, margarine"},"instructions":{"type":"string","example":"pour and frisk egg with margarine"}}},"ModifyRecipe":{"type":"object","properties":{"recipeName":{"type":"string","example":"cake"},"category":{"type":"string","example":"pastries"},"ingredients":{"type":"string","example":"baking powder, flour, margarine"},"instructions":{"type":"string","example":"pour and frisk egg with margarine"}}},"ViewRecipe":{"type":"object","required":["id","userId","views","upvote","downvote","recipeName","ingredients","instructions","createdAt","updatedAt"],"properties":{"id":{"type":"integer","format":"int64","example":1},"recipeName":{"type":"string","example":"cake"},"category":{"type":"string","example":"pastries"},"ingredients":{"type":"string","example":"baking powder, flour, margarine"},"instructions":{"type":"string","example":"pour and frisk egg with margarine"},"views":{"type":"integer","format":"int64","example":2},"upvote":{"type":"integer","format":"int64","example":0},"downvote":{"type":"integer","format":"int64","example":0},"createdAt":{"type":"string","example":"2017-09-19T14:04:22.814Z"},"updatedAt":{"type":"string","example":"2017-09-19T14:04:22.814Z"},"userId":{"type":"integer","format":"int64","example":1}}},"GenericRecipe":{"type":"object","required":["id","views","upvote","downvote","recipeName","ingredients","instructions"],"properties":{"id":{"type":"integer","format":"int64","example":1},"views":{"type":"integer","format":"int64","example":2},"upvote":{"type":"integer","format":"int64","example":12},"downvote":{"type":"integer","format":"int64","example":0},"recipeName":{"type":"string","example":"cake"},"category":{"type":"string","example":"pastries"},"ingredients":{"type":"string","example":"baking powder, flour, margarine"},"instructions":{"type":"string","example":"pour and frisk egg with margarine"}}},"DeleteRecipe":{"type":"object","required":["status","message"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Recipe successfully deleted"}}},"PostedReview":{"type":"object","required":["comment"],"properties":{"comment":{"type":"string","example":"I love this post"}}},"ReviewResponse":{"type":"object","required":["status","message","comment"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Review successfully posted"},"comment":{"type":"string","example":"I love this post"}}},"DownvoteResponse":{"type":"object","required":["status","message","upvote","downvote"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Your vote has been recorded"},"upvote":{"type":"integer","format":"int64","example":2},"downvote":{"type":"integer","format":"int64","example":0}}},"UpvoteResponse":{"type":"object","required":["status","message","upvote","downvote"],"properties":{"status":{"type":"string","example":"success"},"message":{"type":"string","example":"Your vote has been recorded"},"upvote":{"type":"integer","format":"int64","example":1},"downvote":{"type":"integer","format":"int64","example":1}}}}}